% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.R
\name{assign_ids}
\alias{assign_ids}
\title{assign id}
\usage{
assign_ids(x, hid_col, pid_col, keep_id_cols = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{data.frame}

\item{hid_col}{character}

\item{pid_col}{character}
}
\value{
data.frame
}
\description{
assign id
}
