% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{combine_margins}
\alias{combine_margins}
\title{Combine fitted margins and control margins}
\usage{
combine_margins(ml_fit)
}
\arguments{
\item{ml_fit}{a \code{ml_fit} object.}
}
\value{
returns a named list with two components,
\code{individual} and \code{group}. Each contains a list of margins
as \code{data.frame}s.
}
\description{
This function is useful for extracting margins for
further analysis.
}
\examples{
fit <- ml_fit(readRDS(toy_example("Tiny")), "ipu")
combine_margins(fit)
}
