% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_problem.R
\name{fitting_problem}
\alias{fitting_problem}
\alias{format.fitting_problem}
\alias{is.fitting_problem}
\alias{print.fitting_problem}
\alias{special_field_names}
\title{Create an instance of a fitting problem}
\usage{
fitting_problem(ref_sample, controls = list(individual = individual_controls,
  group = group_controls), field_names, individual_controls, group_controls,
  prior_weights = NULL)

is.fitting_problem(x)

\method{format}{fitting_problem}(x, ...)

\method{print}{fitting_problem}(x, ...)

special_field_names(groupId, individualId, individualsPerGroup, count = NULL)
}
\arguments{
\item{ref_sample}{The reference sample}

\item{controls}{Control totals, by default initialized from the
\code{individual_controls} and \code{group_controls} arguments}

\item{field_names}{Names of special fields, construct using
\code{\link{special_field_names}}}

\item{individual_controls, group_controls}{Control totals at individual
and group level, given as a list of data frames where each data frame
defines a control}

\item{prior_weights}{Prior (or design) weights at group level; by default
a vector of ones will be used, which corresponds to random sampling of
groups}

\item{x}{An object}

\item{...}{Ignored.}

\item{groupId, individualId}{Name of the column that defines the ID of the
group or the individual}

\item{individualsPerGroup}{Name of a column that contains the number
of persons in the group}

\item{count}{Name of control total column in control tables (use first numeric
column in each control by default).}
}
\value{
An object of class \code{fitting_problem}, essentially a named list
  with the following components:
\describe{
  \item{\code{refSample}}{The reference sample, a \code{data.frame}.}
  \item{\code{controls}}{A named list with two components, \code{individual}
  and \code{group}. Each contains a list of controls as \code{data.frame}s.}
  \item{\code{fieldNames}}{A named list with the names of special fields.}
}
}
\description{
Create an instance of a fitting problem
}

