% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flatten_ml_fit_problem.R
\name{flatten_ml_fit_problem}
\alias{flatten_ml_fit_problem}
\title{Return a flattened representation of a multi-level fitting problem instance}
\usage{
flatten_ml_fit_problem(ref_sample, controls, field_names, verbose = FALSE)
}
\arguments{
\item{ref_sample}{Reference sample.  Alternatively, a configuration as
returned by \code{\link{import_IPAF_results}}; in this case, the other
arguments are ignored.}

\item{controls}{Control totals}

\item{field_names}{Names of special columns in the data}

\item{verbose}{If \code{TRUE}, print diagnostic output.}
}
\value{
An object of classes \code{flat_ml_fit_problem},
  essentially a named list.
}
\description{
This function transforms a multi-level fitting problem to a representation
more suitable for applying the algorithms:  A matrix with one row per controlled
attribute and one column per unique household, a weight vector with one weight
per household, and a control vector.
}
\examples{
path <- system.file("extdata/minitoy", package="MultiLevelIPF")
flatten_ml_fit_problem(ref_sample = import_IPAF_results(path))
}
\seealso{
\code{\link{ml_fit}}
}

