\name{import_IPAF_results}
\alias{import_IPAF_results}
\title{Imports results from a run of the original Python implementation}
\usage{
import_IPAF_results(path)
}
\arguments{
  \item{path}{Path to the directory that contains the
  input/output files}
}
\value{
Named list
}
\description{
This helper function imports data and results from a run of the original
Python implementation into a named list.
}

