% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_fit.R
\name{ml_fit}
\alias{ml_fit}
\title{Estimate weights using a given algorithm}
\usage{
ml_fit(algorithm = c("entropy_o", "dss"), fitting_problem, verbose = FALSE,
  ...)
}
\arguments{
\item{algorithm}{Algorithm to use}

\item{fitting_problem}{A fitting problem created by
\code{\link{fitting_problem}}.}

\item{verbose}{If \code{TRUE}, print diagnostic output.}

\item{...}{Further parameters passed to the algorithm}
}
\value{
An object of class \code{ml_fit}, essentially a named list.
}
\description{
This function reweights a reference sample to match constraints given by
aggregate controls using an algorithm given as parameter.
}
\examples{
path <- toy_example("minitoy")
ml_fit(algorithm = "entropy_o", fitting_problem = readRDS(path))
}

