% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ml_fit_dss.R
\name{ml_fit_dss}
\alias{ml_fit_dss}
\title{Estimate weights using generalized raking}
\usage{
ml_fit_dss(fitting_problem, method = c("raking", "linear", "logit"),
  ginv = MASS::ginv, verbose = FALSE)
}
\arguments{
\item{fitting_problem}{A fitting problem created by
\code{\link{fitting_problem}} or \code{\link{import_IPAF_results}}.}

\item{method}{a character string specifying the calibration method to be
used.  Possible values are \code{"linear"} for the linear method,
\code{"raking"} for the multiplicative method known as raking and
\code{"logit"} for the logit method.}

\item{ginv}{a function that computes the Moore-Penrose generalized
inverse (default: \code{\link[MASS]{ginv}}). In some cases it is possible to
use a function that computes a "regular" matrix inverse such as \code{{solve.default}}.}

\item{verbose}{If \code{TRUE}, print diagnostic output.}
}
\value{
An object of classes \code{ml_fit_dss} and \code{ml_fit},
  essentially a named list.
}
\description{
This function reweights a reference sample to match constraints given by
aggregate controls by means of generalized raking.
}
\details{
Internally, \code{grake::\link[grake]{calibWeights}} is called.
}
\examples{
path <- system.file("extdata/minitoy", package="MultiLevelIPF")
ml_fit_dss(ref_sample = import_IPAF_results(path), ginv = solve)
}
\references{
Deville, J.-C. and \enc{Särndal}{Saerndal}, C.-E. (1992)
Calibration estimators in survey sampling. \emph{Journal of the American
Statistical Association}, \bold{87}(418), 376--382.

Deville, J.-C., \enc{Särndal}{Saerndal}, C.-E. and Sautory, O. (1993)
Generalized raking procedures in survey sampling. \emph{Journal of the
American Statistical Association}, \bold{88}(423), 1013--1020.
}
\seealso{
\code{\link[grake]{calibWeights}}
}

