% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ml_fit_entropy_o.R
\name{ml_fit_entropy_o}
\alias{ml_fit_entropy_o}
\title{Estimate weights using an entropy optimization approach}
\usage{
ml_fit_entropy_o(fitting_problem, verbose = FALSE, dfsane_args = list())
}
\arguments{
\item{verbose}{If \code{TRUE}, print diagnostic output.}

\item{dfsane_args}{Additional arguments (as a named list) passed to the
\code{\link[BB]{dfsane}} function used internally for the optimization.}
}
\value{
An object of classes \code{ml_fit_entropy_o} and \code{ml_fit},
  essentially a named list.
}
\description{
This function reweights a reference sample to match constraints given by
aggregate controls using an entropy optimization method.
}
\examples{
path <- system.file("extdata/minitoy", package="MultiLevelIPF")
ml_fit_entropy_o(ref_sample = import_IPAF_results(path))
}
\seealso{
\code{\link[BB]{dfsane}}
}

