% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_fit_hipf.R
\name{ml_fit_hipf}
\alias{ml_fit_hipf}
\title{Estimate weights using Hierarchical Iterative Proportional Fitting}
\usage{
ml_fit_hipf(fitting_problem, diff_tol = 16 * .Machine$double.eps,
  tol = 1e-06, maxiter = 200, verbose = FALSE)
}
\arguments{
\item{fitting_problem}{A fitting problem created by
\code{\link{fitting_problem}}.}

\item{diff_tol}{Tolerance, the algorithm stops when relative difference of
control values between iterations drops below this value}

\item{tol}{Tolerance, the algorithm has succeeded when all target values are
reached within this tolerance.}

\item{maxiter}{Maximum number of iterations.}

\item{verbose}{If \code{TRUE}, print diagnostic output.}
}
\value{
An object of classes \code{ml_fit_hipf} and \code{ml_fit}.
}
\description{
This function reweights a reference sample to match constraints given by
aggregate controls by means of Hierarchical Iterative Proportional Fitting.
}
\examples{
path <- toy_example("minitoy")
ml_fit_hipf(fitting_problem = readRDS(path))
}
\references{
Müller, K. and Axhausen, K. W. (2011), Hierarchical IPF: Generating a
       synthetic population for Switzerland, paper presented at the 51st
       Congress of the European Regional Science Association, University of
       Barcelona, Barcelona.
}

