% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_replicate.R
\name{ml_replicate}
\alias{ml_replicate}
\title{Replicate records in a reference sample based on its fitted weights}
\usage{
ml_replicate(algorithm = c("pp", "trs", "round"), ml_fit, verbose = FALSE, ...)
}
\arguments{
\item{algorithm}{Replication algorithm to use}

\item{ml_fit}{A \code{ml_fit} object created by the \code{\link[=ml_fit]{ml_fit()}} family.}

\item{verbose}{If \code{TRUE}, print diagnostic output.}

\item{...}{Further parameters passed to the algorithm}
}
\value{
All functions return a data.frame.
}
\description{
These functions replicate each entry in a reference sample based on its fitted weights.

\code{ml_replicate()} accepts a replication method as argument and calls the
corresponding function. This is useful if the result of multiple
replication algorithms are compared to each other, or to generate a full synthetic
population based on the result of a \code{ml_fit} object.
}
\examples{
path <- toy_example("Tiny")
fit <- ml_fit(algorithm = "entropy_o", fitting_problem = readRDS(path)) 
syn_pop <- ml_replicate(algorithm = "trs", fit)
}
