% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlfit-package.R
\docType{package}
\name{mlfit-package}
\alias{mlfit}
\alias{mlfit-package}
\title{mlfit: Implementation of algorithms that extend Iterative Proportional Fitting to nested structures}
\description{
The Iterative Proportional Fitting (IPF) algorithm operates on count data. This package offers
    implementations for several algorithms that extend this to nested structures:
    "parent" and "child" items for both of which constraints can be provided.
}
\details{
To use this package, you need to:
\enumerate{
\item Specify your fitting problem with \code{\link[=ml_problem]{ml_problem()}}
\item Optionally, convert the fitting problem to a structure that can be
processed by the algorithms with \code{\link[=flatten_ml_fit_problem]{flatten_ml_fit_problem()}}; this is
helpful if you want to run the same fitting problem with multiple
algorithms and compare results.
\item Compute weights with one of the algorithms provided in this package with
\code{\link[=ml_fit]{ml_fit()}} or one of the specialized functions
\item Analyze weights or residuals, e.g. with \code{\link[=compute_margins]{compute_margins()}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://krlmlr.github.io/mlfit}
  \item \url{https://github.com/krlmlr/mlfit}
  \item Report bugs at \url{https://github.com/krlmlr/mlfit/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

Other contributors:
\itemize{
  \item Amarin Siripanich \email{amarin.siri@gmail.com} [contributor]
}

}
