\docType{package}
\name{MultiLevelIPF-package}
\alias{MultiLevelIPF}
\alias{MultiLevelIPF-package}
\title{Implementation of algorithms that extend IPF to nested structures}
\description{
The IPF algorithm operates on count data.  This package offers
implementations for several algorithms that extend this to nested
structures: "parent" and "child" items for both of which constraints can be
provided.
}
\details{
\tabular{ll}{
  Package: \tab MultiLevelIPF\cr
  Type: \tab Package\cr
  Version: \tab 1.0\cr
  Date: \tab 2013-07-09\cr
  License: \tab What license is it under?\cr
} ~~
An overview of how to use the package, including the most
important functions ~~
}
\examples{
# ~~ simple examples of the most important functions ~~
}
\author{
Kirill Müller

Maintainer: Kirill Müller <kirill.mueller@ivt.baug.ethz.ch>

Bug reports: https://github.com/krlmlr/MultiLevelIPF/issues
}
\references{
Müller, K. and Axhausen, K. W. (2011), Hierarchical IPF: Generating a
       synthetic population for Switzerland, in ERSA, ed., ‘51st Congress of
       the European Regional Science Association’, University of Barcelona,
       Barcelona.
}
\seealso{
~~ Optional links to other man pages, e.g. ~~ ~~
<pkg>:<pkg>-package ~~
}
\keyword{package}

