% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy.R
\name{toy_example}
\alias{toy_example}
\title{Access to toy examples bundled in this package}
\usage{
toy_example(name = NULL)
}
\arguments{
\item{name}{Name of the example, default: return all}
}
\value{
A named vector of file system paths.
}
\description{
Returns the paths to all available toy examples, or to a specific toy
example.  Load via \code{\link[=readRDS]{readRDS()}}.
}
\examples{
toy_example()
}
