% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.R
\name{trs}
\alias{trs}
\title{Truncate, Replicate, and Sample}
\usage{
trs(x, w, hid_col, pid_col, verbose = TRUE)
}
\arguments{
\item{x}{a data.frame or ml_fit object}

\item{w}{a numeric vector which has the same length as the rows in \code{x}}

\item{hid_col}{a character indicates household id/group id column}

\item{pid_col}{a character indicates individual id coluin}
}
\value{
a data.frame
}
\description{
For more details about the TRS method see its original paper
Lovelace, R. and Ballas, D., 2013. ‘Truncate, replicate, sample’:
A method for creating integer weights for spatial microsimulation. Computers,
Environment and Urban Systems, 41, pp.1-11. Vancouver
}
\examples{

fp <- readRDS(toy_example("Joint"))
fit <- ml_fit_ipu(fp)
synthetic_population <- trs(fit)
}
