% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/margins.R
\name{compute_margins}
\alias{compute_margins}
\title{Compute margins for a weighting of a multi-level fitting problem}
\usage{
compute_margins(fitting_problem, weights, verbose = FALSE)
}
\arguments{
\item{fitting_problem}{A fitting problem created by
\code{\link{fitting_problem}} or \code{\link{import_IPAF_results}}.}

\item{weights}{A vector with one entry per row of the original reference
sample}

\item{verbose}{If \code{TRUE}, print diagnostic output.}
}
\value{
Margins of the same format as expected by the \code{controls}
  parameter of the \code{\link{fitting_problem}} function.
}
\description{
This function allows checking a fit in terms of the original input data.
}
\seealso{
\code{\link{ml_fit}}
}

