% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_problem.R
\name{fitting_problem}
\alias{fitting_problem}
\title{Create a fitting problem}
\usage{
fitting_problem(
  ref_sample,
  controls = list(individual = individual_controls, group = group_controls),
  field_names,
  individual_controls,
  group_controls,
  prior_weights = NULL
)
}
\value{
See \code{\link[=ml_problem]{ml_problem()}}.
}
\description{
Soft-deprecated, new code should use \code{\link[=ml_problem]{ml_problem()}}.
}
\keyword{internal}
