% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_fit.R, R/ml_fit_dss.R, R/ml_fit_entropy_o.R, R/ml_fit_hipf.R, R/ml_fit_ipu.R
\name{ml_fit}
\alias{ml_fit}
\alias{ml_fit_dss}
\alias{ml_fit_entropy_o}
\alias{ml_fit_hipf}
\alias{ml_fit_ipu}
\title{Estimate weights for a fitting problem}
\usage{
ml_fit(algorithm = c("entropy_o", "dss", "ipu", "hipf"), fitting_problem,
  verbose = FALSE, ..., tol = 1e-06)

ml_fit_dss(fitting_problem, method = c("raking", "linear", "logit"),
  ginv = grake::gginv(), tol = 1e-06, verbose = FALSE)

ml_fit_entropy_o(fitting_problem, verbose = FALSE, tol = 1e-06,
  dfsane_args = list())

ml_fit_hipf(fitting_problem, diff_tol = 16 * .Machine$double.eps,
  tol = 1e-06, maxiter = 200, verbose = FALSE)

ml_fit_ipu(fitting_problem, diff_tol = 16 * .Machine$double.eps,
  tol = 1e-06, maxiter = 2000, verbose = FALSE)
}
\arguments{
\item{algorithm}{Algorithm to use}

\item{fitting_problem}{A fitting problem created by
\code{\link{fitting_problem}}.}

\item{verbose}{If \code{TRUE}, print diagnostic output.}

\item{...}{Further parameters passed to the algorithm}

\item{tol}{Tolerance, the algorithm has succeeded when all target values are
reached within this tolerance.}

\item{method}{Calibration method, one of \code{"raking"} (default),
\code{"linear"}, or \code{"logit"}}

\item{ginv}{Function that computes the Moore-Penrose pseudoinverse.}

\item{dfsane_args}{Additional arguments (as a named list) passed to the
\code{\link[BB]{dfsane}} function used internally for the optimization.}

\item{diff_tol}{Tolerance, the algorithm stops when relative difference of
control values between iterations drops below this value}

\item{maxiter}{Maximum number of iterations.}
}
\value{
All functions return an object of class \code{ml_fit}, which is
  a named list under the hood.  The class matches the function called,
  e.g., the return value of the \code{ml_fit_ipu} function also is of class
  \code{ml_fit_ipu}.
}
\description{
These functions reweight a reference sample to match constraints given by
  aggregate controls.

\code{ml_fit} accepts an algorithm as argument and calls the
  corresponding function. This is useful if the result of multiple algorithms
  are compared to each other.

\code{ml_fit_dss} uses generalized raking and calls \code{grake::\link[grake]{dss}} internally.

\code{ml_fit_entropy_o} optimizes entropy using the
  \code{BB::\link[BB]{dfsane}} function.

\code{ml_fit_hipf} implements Hierarchical Iterative Proportional Fitting.

\code{ml_fit_ipu} implements Iterative Proportional Updating.
}
\examples{
path <- toy_example("Tiny")
ml_fit(algorithm = "entropy_o", fitting_problem = readRDS(path))
ml_fit_dss(fitting_problem = readRDS(path))
ml_fit_dss(fitting_problem = readRDS(path), ginv = solve)
ml_fit_entropy_o(fitting_problem = readRDS(path))
ml_fit_hipf(fitting_problem = readRDS(path))
ml_fit_ipu(fitting_problem = readRDS(path))
}
\references{
Deville, J.-C. and \enc{Särndal}{Saerndal}, C.-E. (1992)
Calibration estimators in survey sampling. \emph{Journal of the American
Statistical Association}, \bold{87} (418), 376--382.

Deville, J.-C., \enc{Särndal}{Saerndal}, C.-E. and Sautory, O. (1993)
Generalized raking procedures in survey sampling. \emph{Journal of the
American Statistical Association}, \bold{88} (423), 1013--1020.

Müller, K. and Axhausen, K. W. (2011), Hierarchical IPF: Generating a
       synthetic population for Switzerland, paper presented at the 51st
       Congress of the European Regional Science Association, University of
       Barcelona, Barcelona.

Ye, X., K. Konduri, R. M. Pendyala, B. Sana and P. A. Waddell (2009)
A methodology to match distributions of both household and person attributes
in the generation of synthetic populations, paper presented at the \emph{88th
Annual Meeting of the Transportation Research Board}, Washington, D.C.,
January 2009.
}
\seealso{
\code{\link[grake]{dss}}, \code{\link[grake]{gginv}}

\code{\link[BB]{dfsane}}
}

