% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_fit_dss.R
\name{ml_fit_dss}
\alias{ml_fit_dss}
\title{Estimate weights using generalized raking}
\usage{
ml_fit_dss(fitting_problem, method = c("raking", "linear", "logit"),
  ginv = grake::gginv(), tol = 1e-06, verbose = FALSE)
}
\arguments{
\item{fitting_problem}{A fitting problem created by
\code{\link{fitting_problem}}.}

\item{method}{Calibration method, one of \code{"raking"} (default),
\code{"linear"}, or \code{"logit"}}

\item{ginv}{Function that computes the Moore-Penrose pseudoinverse,
currently only the default value \code{MASS::\link[MASS]{ginv}} is
accepted.}

\item{tol}{Tolerance, the algorithm has succeeded when all target values are
reached within this tolerance.}

\item{verbose}{If \code{TRUE}, print diagnostic output.}
}
\value{
An object of classes \code{ml_fit_dss} and \code{ml_fit},
  essentially a named list.
}
\description{
This function reweights a reference sample to match constraints given by
aggregate controls by means of generalized raking.
}
\details{
Internally, \code{grake::\link[grake]{dss}} is called.
}
\examples{
path <- toy_example("minitoy")
ml_fit_dss(fitting_problem = readRDS(path))
\dontrun{ml_fit_dss(fitting_problem = readRDS(path), ginv = solve)}
}
\references{
Deville, J.-C. and \enc{Särndal}{Saerndal}, C.-E. (1992)
Calibration estimators in survey sampling. \emph{Journal of the American
Statistical Association}, \bold{87}(418), 376--382.

Deville, J.-C., \enc{Särndal}{Saerndal}, C.-E. and Sautory, O. (1993)
Generalized raking procedures in survey sampling. \emph{Journal of the
American Statistical Association}, \bold{88}(423), 1013--1020.
}
\seealso{
\code{\link[grake]{dss}}
}

