% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ml_fit_dss.R
\name{ml_fit_dss}
\alias{ml_fit_dss}
\title{Estimate weights using generalized raking}
\usage{
ml_fit_dss(ref_sample, controls, field_names, method = c("raking", "linear",
  "logit"), verbose = FALSE)
}
\arguments{
\item{ref_sample}{Reference sample.  Alternatively, a configuration as
returned by \code{\link{import_IPAF_results}}; in this case, the other
arguments are ignored.}

\item{controls}{Control totals}

\item{field_names}{Names of special columns in the data}

\item{method}{a character string specifying the calibration method to be
used.  Possible values are \code{"linear"} for the linear method,
\code{"raking"} for the multiplicative method known as raking and
\code{"logit"} for the logit method.}

\item{verbose}{If \code{TRUE}, print diagnostic output.}
}
\value{
An object of classes \code{ml_fit_dss} and \code{ml_fit},
  essentially a named list.
}
\description{
This function reweights a reference sample to match constraints given by
aggregate controls by means of generalized raking.
}
\details{
Internally, \code{laeken::\link[laeken]{calibWeights}} is called.
}
\examples{
path <- system.file("extdata/minitoy", package="MultiLevelIPF")
ml_fit_dss(ref_sample = import_IPAF_results(path))
}
\references{
Deville, J.-C. and \enc{Särndal}{Saerndal}, C.-E. (1992)
Calibration estimators in survey sampling. \emph{Journal of the American
Statistical Association}, \bold{87}(418), 376--382.

Deville, J.-C., \enc{Särndal}{Saerndal}, C.-E. and Sautory, O. (1993)
Generalized raking procedures in survey sampling. \emph{Journal of the
American Statistical Association}, \bold{88}(423), 1013--1020.
}
\seealso{
\code{\link[laeken]{calibWeights}}
}

