% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ml_fit_entropy_o}
\alias{ml_fit_entropy_o}
\title{Estimate weights using an entropy optimization approach}
\usage{
ml_fit_entropy_o(ref_sample, controls, field_names, verbose = FALSE,
  dfsane_args = list())
}
\arguments{
\item{ref_sample}{Reference sample.  Alternatively, a configuration as
returned by \code{\link{import_IPAF_results}}; in this case, the other
arguments are ignored.}

\item{controls}{Control totals}

\item{field_names}{Names of special columns in the data}

\item{verbose}{If \code{TRUE}, print diagnostic output.}

\item{dfsane_args}{Additional arguments (as a named list) passed to the
\code{\link[BB]{dfsane}} function used internally for the optimization.}
}
\value{
An object of classes \code{ml_fit_entropy_o} and \code{ml_fit},
  essentially a named list.
}
\description{
This function reweights a reference sample to match constraints given by
aggregate controls using an entropy optimization method.
}
\examples{
path <- system.file("extdata/minitoy", package="MultiLevelIPF")
ml_fit_entropy_o(ref_sample = import_IPAF_results(path))
}
\seealso{
\code{\link[BB]{dfsane}}
}

