% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_fit_entropy_o.R
\name{ml_fit_entropy_o}
\alias{ml_fit_entropy_o}
\title{Estimate weights using an entropy optimization approach}
\usage{
ml_fit_entropy_o(fitting_problem, verbose = FALSE, tol = 1e-06,
  dfsane_args = list())
}
\arguments{
\item{fitting_problem}{A fitting problem created by
\code{\link{fitting_problem}}.}

\item{verbose}{If \code{TRUE}, print diagnostic output.}

\item{tol}{Tolerance, the algorithm has succeeded when all target values are
reached within this tolerance.}

\item{dfsane_args}{Additional arguments (as a named list) passed to the
\code{\link[BB]{dfsane}} function used internally for the optimization.}
}
\value{
An object of classes \code{ml_fit_entropy_o} and \code{ml_fit},
  essentially a named list.
}
\description{
This function reweights a reference sample to match constraints given by
aggregate controls using an entropy optimization method.
}
\examples{
path <- toy_example("minitoy")
ml_fit_entropy_o(fitting_problem = readRDS(path))
}
\seealso{
\code{\link[BB]{dfsane}}
}

