% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_fit_ipu.R
\name{ml_fit_ipu}
\alias{ml_fit_ipu}
\title{Estimate weights using Iterative Proportional Updating}
\usage{
ml_fit_ipu(fitting_problem, tol = 1e-06, maxiter = 5000, verbose = FALSE)
}
\arguments{
\item{fitting_problem}{A fitting problem created by
\code{\link{fitting_problem}}.}

\item{verbose}{If \code{TRUE}, print diagnostic output.}
}
\value{
An object of classes \code{ml_fit_ipu} and \code{ml_fit}.
}
\description{
This function reweights a reference sample to match constraints given by
aggregate controls by means of Iterative Proportional Updating.
}
\examples{
path <- toy_example("minitoy")
ml_fit_ipu(fitting_problem = readRDS(path))
\dontrun{ml_fit_ipu(fitting_problem = readRDS(path), ginv = solve)}
}
\references{
Ye, X., K. Konduri, R. M. Pendyala, B. Sana and P. A. Waddell (2009)
A methodology to match distributions of both household and person attributes
in the generation of synthetic populations, paper presented at the \emph{88th
Annual Meeting of the Transportation Research Board}, Washington, D.C.,
January 2009.
}

