% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiLevelIPF-package.R
\docType{package}
\name{MultiLevelIPF-package}
\alias{MultiLevelIPF}
\alias{MultiLevelIPF-package}
\title{Implementation of algorithms that extend IPF to nested structures}
\description{
The IPF algorithm operates on count data.  This package offers
implementations for several algorithms that extend this to nested
structures: "parent" and "child" items for both of which constraints can be
provided.
}
\examples{
\dontrun{import_IPAF_results("path/to/results")}
}
\references{
Müller, K. and Axhausen, K. W. (2011), Hierarchical IPF: Generating a
       synthetic population for Switzerland, paper presented at the 51st
       Congress of the European Regional Science Association, University of
       Barcelona, Barcelona.
}
\keyword{package}

