\docType{package}
\name{MultiLevelIPF-package}
\alias{MultiLevelIPF}
\alias{MultiLevelIPF-package}
\title{Implementation of algorithms that extend IPF to nested structures}
\description{
The IPF algorithm operates on count data.  This package offers
implementations for several algorithms that extend this to nested
structures: "parent" and "child" items for both of which constraints can be
provided.
}
\details{
\tabular{ll}{
  Package: \tab MultiLevelIPF\cr
  Type: \tab Package\cr
  Version: \tab 0.0.6\cr
  Date: \tab 2014-02-09\cr
  License: \tab GPL v3\cr
} ~~
An overview of how to use the package, including the most
important functions ~~
}
\examples{
\dontrun{import_IPAF_results("path/to/results")}
}
\author{
Kirill Müller

Maintainer: Kirill Müller <\href{mailto:kirill.mueller@ivt.baug.ethz.ch}{kirill.mueller@ivt.baug.ethz.ch}>

Bug reports: \url{https://github.com/krlmlr/MultiLevelIPF/issues}
}
\references{
Müller, K. and Axhausen, K. W. (2011), Hierarchical IPF: Generating a
       synthetic population for Switzerland, paper presented at the 51st
       Congress of the European Regional Science Association, University of
       Barcelona, Barcelona.
}
\seealso{
~~ Optional links to other man pages, e.g. ~~ ~~
<pkg>:<pkg>-package ~~
}
\keyword{package}

